[feat]: add Kotlin DAO example

+++ firefly-example/src/main/kotlin/com/firefly/example/kotlin/coffee/store/dao/ProductDAO.kt
@@ -17,6 +17,6 @@ interface ProductDAO {
 
     suspend fun insert(product: Product): Long
 
-    suspend fun update(product: Product): Long
+    suspend fun update(product: Product): Int
 
 }
